\name{RasterPatch-class}

\docType{class}

\alias{RasterPatch-class}

\alias{show,RasterPatch-method}


\title{ RasterPatch class}

\description{
describe this class
}


\section{Objects from the Class}{
Objects can be created by 
}

\section{Slots}{
Slots for Raster* objects
\describe{
   \item{\code{patch}:}{Object of class .Patch} 
}
}

\examples{
showClass("RasterPatch")
}

\keyword{classes}
\keyword{spatial}
