\name{fragstats}


\alias{fragstats}

\title{fragstats}

\description{
Wrapper around fragstats 2.0 code  -- experimental
}

\usage{
fragstats(...)
}

\arguments{
  \item{...}{nothing so far}
}

\details{
none 
}

\value{
NULL
}

\author{Robert J. Hijmans}

\examples{
fragstats()
}

\keyword{spatial}

