\name{fragstats}


\alias{fragstats}

\title{fragstats}

\description{
Wrapper around fragstats 2.0 code  -- experimental
}

\usage{
fragstats(...)
}

\arguments{
  \item{...}{nothing so far}
}

\details{
This is a wrapper to FRAGSTATS version 2, that was released in the public domain in 1995 and described in 
\href{http://www.umass.edu/landeco/pubs/mcgarigal.marks.1995.pdf}{McGarigal and Marks 1995}. See this \href{http://andrewsforest.oregonstate.edu/data/tools/software/fragstat.cfm?topnav=149}{link} for info.
}

\value{
NULL
}

\author{Robert J. Hijmans}

\examples{
fragstats()
}

\keyword{spatial}

